Implements chat video preview.
